:recycle: Refactor HF trainer logic and move to utils
