+++ caikit_nlp/modules/text_generation/text_generation_local.py
@@ -48,6 +48,7 @@ from ...toolkit.text_generation.model_run_utils import (
     generate_text_func,
 )
 from ...toolkit.torch_run import get_torch_elastic_launch_config
+from ...toolkit.text_generation.training_utils import preprocess_function, launch_training, infer_max_steps
 
 log = alog.use_channel("TXT_GEN")
 error = error_handler.get(log)
@@ -284,7 +285,7 @@ class TextGeneration(ModuleBase):
         ## Generate data loader from stream
         training_dataset: Union[
             Dataset, TransformersIterableDataset
-        ] = cls._preprocess_function(
+        ] = preprocess_function(
             base_model=base_model,
             train_stream=train_stream,
             tokenizer=base_model.tokenizer,
@@ -292,6 +293,7 @@ class TextGeneration(ModuleBase):
             max_target_length=max_target_length,
             shuffle=True,
             use_iterable_dataset=use_iterable_dataset,
+            random_seed=cls.RANDOM_SEED
         )
 
         ### Dtype based processing
@@ -372,7 +374,7 @@ class TextGeneration(ModuleBase):
                 # negatively impact the performance
                 "full_determinism": False,
                 # Required for iterable dataset
-                "max_steps": cls.infer_max_steps(
+                "max_steps": infer_max_steps(
                     num_epochs, batch_size, training_dataset
                 ),
                 # Some interesting parameters:
@@ -405,7 +407,7 @@ class TextGeneration(ModuleBase):
                 # to avoid that, specially for unit tests, we are only
                 # running below when GPUs are available
                 training_loss_history = torch.distributed.launcher.api.elastic_launch(
-                    launch_config, cls._launch_training
+                    launch_config, launch_training
                 )(base_model, training_dataset, training_args, checkpoint_dir)
 
                 # NOTE: We are currently only storing the loss information from
@@ -413,7 +415,7 @@ class TextGeneration(ModuleBase):
                 # rank of the process as key
                 training_loss_history = training_loss_history[0]
             else:
-                training_loss_history = cls._launch_training(
+                training_loss_history = launch_training(
                     base_model, training_dataset, training_args, checkpoint_dir
                 )
 
@@ -566,98 +568,3 @@ class TextGeneration(ModuleBase):
             max_time=max_time,
             **kwargs,
         )
-
-    ################################## Private Functions ######################################
-
-    @staticmethod
-    def _preprocess_function(
-        base_model: PretrainedModelBase,
-        train_stream: DataStream[GenerationTrainRecord],
-        tokenizer: AutoTokenizer,
-        max_source_length: int,
-        max_target_length: int,
-        shuffle: bool,
-        use_iterable_dataset: bool,
-    ):
-        """Pre-process each example to get it prepared for training."""
-        dataset_type = TransformersIterableDataset if use_iterable_dataset else Dataset
-        log.debug("Loading dataset class: [%s]", dataset_type.__name__)
-        fn_kwargs = {
-            "tokenizer": tokenizer,
-            "max_source_length": max_source_length,
-            "max_target_length": max_target_length,
-        }
-        dataset = dataset_type.from_generator(
-            get, gen_kwargs={"train_stream": train_stream}
-        )
-        mapped_dataset = dataset.map(
-            base_model.tokenize_function,
-            fn_kwargs=fn_kwargs,
-            batched=base_model.REQUIRES_TOKEN_UNWRAPPING,
-            # Drop the input / output columns; we need to do this for dimensions to play
-            # happily when operating on batched inputs for causal language modeling.
-            remove_columns=["input", "output"],
-        )
-
-        if shuffle:
-            log.debug("Shuffling the dataset")
-            return mapped_dataset.shuffle(seed=TextGeneration.RANDOM_SEED)
-
-        return mapped_dataset
-
-    @staticmethod
-    def _launch_training(
-        base_model, training_dataset, training_args, checkpoint_dir
-    ) -> None:
-        """Utility function to wrap trainer and execute training"""
-
-        trainer = base_model.get_trainer(
-            train_dataset=training_dataset, **training_args
-        )
-
-        # Start training via Trainer.train function
-        trainer.train()
-
-        # save the model temporarily and reload it
-        # this is done, since otherwise the model might be distributed in different
-        # devices, in which case its better to use trainer's `prediction_step`
-        # functions, but then, they don't always give API similar to `generate`
-        # and thus cause incompatibilities in `run` function
-        trainer.save_state()
-        trainer.save_model(checkpoint_dir)
-
-        # save tokenizer explicitly
-        base_model.tokenizer.save_pretrained(checkpoint_dir)
-
-        # Below will return log history but launch will automatically attach rank to it.
-        # if started in distributed fashion
-        return trainer.state.log_history
-
-    @staticmethod
-    def infer_max_steps(
-        num_epochs: int,
-        batch_size: int,
-        training_dataset: Union[Dataset, TransformersIterableDataset],
-    ):
-        # Calculate the number of samples that we have
-        if isinstance(training_dataset, Dataset):
-            data_len = len(training_dataset)
-        else:
-            data_len = 0
-            for _ in training_dataset:
-                data_len += 1
-        # Figure out how many batches we'll have per epoch
-        num_batches = data_len // batch_size
-        # Assume drop_last=False; in general, this doesn't really matter.
-        # We mostly do this to avoid strange behavior when the dataset
-        # size is smaller than the batch size.
-        if num_batches != (data_len * batch_size):
-            num_batches += 1
-        num_steps = num_batches * num_epochs
-        log.debug("Number of inferred steps: [%s]", num_steps)
-        return num_steps
-
-
-def get(train_stream):
-    for data in train_stream:
-        yield {"input": data.input, "output": data.output}
